+++ _layouts/default.html
@@ -29,7 +29,7 @@
     </main>
 
     {%- include footer.html -%}
-    <script src="/startups/assets/js/termo.js"></script>
+    <script src="{{ '/startups/assets/js/bundle.js' | relative_url }}"></script>
 
   </div>
 </body>
